fix(combobox): renamed autoSelect as autoFilter
