+++ packages/components/combobox/src/Combobox.doc.mdx
@@ -102,13 +102,13 @@ Use `disabled` on individual `Combobox.Item` to disable them.
 
 <Canvas of={stories.DisabledItem} />
 
-### Filtering - AutoSelect
+### Filtering - AutoFilter
 
-Use `autoSelect` to filter out items that does not match the input value. This behaviour is not case-sensitive.
+Use `autoFilter` to filter out items that does not match the input value. This behaviour is not case-sensitive.
 
 For more custom filtering, logic should be done on call-site to render only desired items.
 
-<Canvas of={stories.FilteringAutoSelect} />
+<Canvas of={stories.FilteringAutoFilter} />
 
 ### Filtering - Manual
